% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSpliceJunctionTable.R
\name{CreateSpliceJunctionTable}
\alias{CreateSpliceJunctionTable}
\title{Create a table of unique splice junction expression per tab file (*.out.tab) generated using during
some alignment processes (e.g. using the STAR aligner).}
\usage{
CreateSpliceJunctionTable(tabfiles)
}
\arguments{
\item{tabfiles}{directory path of *.out.tab files generated during alignment.}
}
\value{
A data.frame of the unique splice junction expression per tab file (*.out.tab).
}
\description{
Create a table of unique splice junction expression per tab file (*.out.tab) generated using during
some alignment processes (e.g. using the STAR aligner).
}
\examples{
tabfiles <- dir(path="tabs/", pattern=".*.out.tab$", full.names=T)
JunctionTables <- CreateSpliceJunctionTable(tabfiles)
}
\author{
Nathaniel J. Madrid, Jason Byars
}
\keyword{CreateSpliceJunctionTable}
