% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAlignmentStatistics.R
\name{GetAlignmentStatistics}
\alias{GetAlignmentStatistics}
\title{Get a summary of alignment statistics for an aligned bam file as defined by samtools idxstats.}
\usage{
GetAlignmentStatistics(BamFilePath)
}
\arguments{
\item{BamFilePath}{location of aligned bam file. Corresponding bai index file must be in the same directory.}
}
\value{
A data.frame of a summary of alignment statistics for an aligned bam file as defined by samtools idxstats.
}
\description{
Get a summary of alignment statistics for an aligned bam file as defined by samtools idxstats.
}
\examples{
BamFilePaths <- dir(path="/home/ubuntu/Projects/RNAseqPackage/bams", pattern=".*.bam$", full.names=T)
AlignmentStatsDF = GetAlignmentStatistics(BamFilePaths[1])
}
\author{
Nathaniel J. Madrid, Jason Byars
}
\keyword{GetAlignmentStatistics}
