% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBedGraphRefSeqOverlaps.R
\name{GetBedGraphRefSeqOverlaps}
\alias{GetBedGraphRefSeqOverlaps}
\title{Get overlaps using GenomicRanges findOverlaps for a bedgraph coverage file (*.bedgraph.gz) compared to RefSeq defined gene regions.}
\usage{
GetBedGraphRefSeqOverlaps(bedgraph, refseq)
}
\arguments{
\item{bedgraph}{GRanges of bedgraph coverage file (*.bedgraph.gz) created via bedtools genomecov -bga -split.}

\item{refseq}{GRanges of refseq defined gene regions.}
}
\value{
Results are returned as a Hits object.
}
\description{
Get overlaps using GenomicRanges findOverlaps for a bedgraph coverage file (*.bedgraph.gz) compared to RefSeq defined gene regions.
}
\examples{
refseq <- import.gff("RefSeq_hg19_exons_nodups_021114g1k.gff")
refseq$gene <- sub("_exon_[0-9]+", "", refseq$gene_id)
refseq$width <- width(refseq)
txlength <- tapply(refseq$width, refseq$gene, sum)
# Get file paths of aligned bedgraph files
# Coverage calculated w/ bedtools genomecov -bga -split
BedGraphFiles <- dir(path="/home/bedgraphs", full.names=T)
BedgraphList <- lapply(BedGraphFiles[1:3], function(bg) import.bedGraph(bg))
BedGraphRefSeqOverlapsList <- lapply(BedgraphList, function(bg) GetBedGraphRefSeqOverlaps(bg, refseq))
}
\author{
Nathaniel J. Madrid, Jason Byars
}
\keyword{GetBedGraphRefSeqOverlaps}
