% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCountsPerRegion.R
\name{GetCountsPerRegion}
\alias{GetCountsPerRegion}
\title{Count the number of reads that are in a given genomic region.}
\usage{
GetCountsPerRegion(BamFilePath, param)
}
\arguments{
\item{BamFilePath}{location of aligned bam file. Corresponding bai index file must be in the same directory.}

\item{param}{ScanBamParam data structure.}
}
\value{
An integer value for the number of reads that are in a given genomic region.
}
\description{
Count the number of reads that are in a given genomic region.
}
\examples{
# Get file paths of aligned bam files. The corresponding bai index file must also be in the same directory.
BamFilePaths <- dir(path="/home/BamFiles", pattern=".*.bam$", full.names=T)
rRNAparamRegions <- GRanges(seqnames="GL000220.1", IRanges(1, 161802))

# Note that any region(s) can be used in GetCountsPerRegion, not just rRNA regions. 
rRNAparam <- ScanBamParam(which=rRNAparamRegions)

# Secondary alignments will only be counted once
rRNATotalCounts <- GetCountsPerRegion(BamFilePaths[1], rRNAparam)
}
\author{
Nathaniel J. Madrid, Jason Byars
}
\keyword{GetCountsPerRegion}
