% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHtseqSummaryStats.R
\name{GetHtseqSummaryStats}
\alias{GetHtseqSummaryStats}
\title{Read and organize htseq *_summary_stats.txt files into one data structure.}
\usage{
GetHtseqSummaryStats(HtseqFilePaths)
}
\arguments{
\item{HtseqFilePaths}{htseq *_summary_stats.txt file paths.}
}
\value{
A data.frame containing the summary statistics from each HtSeq file in HtseqFilePaths.
}
\description{
Read and organize htseq *_summary_stats.txt files into one data structure.
}
\examples{
HtseqFilePaths <- dir(path=".", pattern=".*_summary_stats.txt", full.names=T)
HtseqStatsPerFile <- GetHtseqSummaryStats(HtseqFilePaths)
}
\author{
Nathaniel J. Madrid, Jason Byars
}
\keyword{GetHtseqSummaryStats}
