% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumGenesExpressed.R
\name{GetNumGenesExpressed}
\alias{GetNumGenesExpressed}
\title{Get the total number of genes expressed per gtf file (generated via Cufflinks).}
\usage{
GetNumGenesExpressed(gtfDFs, covThreshold = 1)
}
\arguments{
\item{gtfDFs}{data table structure containing Cufflinks results.}

\item{covThreshold}{only genes will be count that have "cov" values above this threshold. Default value is 1.0.}
}
\value{
An integer for the total number of genes expressed per gtf file (generated via Cufflinks).
}
\description{
Get the total number of genes expressed per gtf file (generated via Cufflinks).
}
\examples{
gtffiles <- c("transcripts_sorted.gtf")
FileCons <- mclapply(gtffiles, function(FileName) file(FileName), mc.cores=8)
gtfs <- mclapply(FileCons, function(fc) import.gff(fc, format="gtf"), mc.cores=8)
for(i in length(FileCons)) { close(FileCons[[i]]) }
gtfDFs <- mclapply(gtfs, function(gtf) as.data.frame(gtf[gtf$type=="transcript"]), mc.cores=8)
names(gtfDFs) <- basename(gtffiles)
NumGenesPerFile <- GetNumGenesExpressed(gtfDFs, 1.0)
}
\author{
Nathaniel J. Madrid, Jason Byars
}
\keyword{GetNumGenesExpressed}
