% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPileUpPerGene.R
\name{GetPileUpPerGene}
\alias{GetPileUpPerGene}
\title{Get read depths per gene as defined by the RSamtools pileup function.}
\usage{
GetPileUpPerGene(BamFilePath, TopCDSDF, max_depth = 10000)
}
\arguments{
\item{BamFilePath}{Location of aligned bam file. Corresponding bai index file must be in the same directory.}

\item{TopCDSDF}{Data frame of CDS returned from select(Homo.sapiens).}

\item{max_depth}{Parameter of the same name passed to Rsamtools pileup. Default value is 10000.}
}
\value{
A data.frame created via Rsamtools:pileup for each CDS region in TopCDSDF.
}
\description{
Get read depths per gene as defined by the RSamtools pileup function.
}
\examples{
BamFilePaths <- dir(path="/home/ubuntu/BamFiles", pattern=".*.bam$", full.names=T)
TopGenesDF <- data.frame(gene = c("NM_016824", "NM_019903", "NP_001112"))
cdsColumns <- c("CDSCHROM", "CDSSTART", "CDSEND", "CDSSTRAND", "SYMBOL")
TopCDSDF <- select(Homo.sapiens, keys=as.character(TopGenesDF$gene), columns=cdsColumns, keytype="REFSEQ")
TopCDSDF$CDSCHROM <- sub("chr", "", TopCDSDF$CDSCHROM)
TopCDSDF <- TopCDSDF[which(TopCDSDF$CDSCHROM \%in\% c(1:22, "X", "Y")), ]
PileUpPerFile <- lapply(BamFilePaths[1:3], function(f) GetPileUpPerGene(f, TopCDSDF)) 
}
\author{
Nathaniel J. Madrid, Jason Byars
}
\keyword{GetPileUpPerGene}
