% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetrRNAStatistics.R
\name{GetrRNAStatistics}
\alias{GetrRNAStatistics}
\title{Get the total number of rRNA read counts from bam file.
samtools idxstats must be installed on the local system.}
\usage{
GetrRNAStatistics(BamFilePath)
}
\arguments{
\item{BamFilePath}{location of aligned bam file. Corresponding bai index file must be in the same directory.}
}
\value{
An integer for the total number of rRNA read counts in a given bam file.
}
\description{
Note that samtools idxstats may or may not include secondary alignments. This will depend on how the reads
were aligned. See the documention for samtools idxstats and the aligner used to create the bam file.
The bam file must have been aligned to a reference genome containing chromosone GL000220.1
total reads mapping to GL0000220.1 or know rRNA regions defined by Ensembl
}
\examples{
BamFilePaths <- dir(path="/home/ubuntu/Projects/RNAseqPackage/bams", pattern=".*.bam$", full.names=T)
# samtools idxstats must be installed on the local system.
# Note that samtools idxstats may or may not include secondary alignments. This will depend on how the reads
# were aligned. See the documention for samtools idxstats and the aligner used to create the bam file.
rRNAStatsDF = GetrRNAStatistics(BamFilePaths[1])
}
\author{
Nathaniel J. Madrid, Jason Byars
}
\keyword{GetrRNAStatistics}
