% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeUniqueReads.R
\name{SummarizeUniqueReads}
\alias{SummarizeUniqueReads}
\title{Get a summary of the unique reads from bam file. Duplicate reads are defined as having identical pos and cigar values.}
\usage{
SummarizeUniqueReads(BamFilePath)
}
\arguments{
\item{BamFilePath}{location of aligned bam file. Corresponding bai index file must be in the same directory.}
}
\value{
A data.frame containing a summary of the unique reads from bam file. Duplicate reads are defined as having identical pos and cigar values.
}
\description{
unique aligned reads = total number of reads that are not secondary;
uniquely aligned reads without duplicates = total number of reads that are not secondary and not duplicates;
duplicate reads = total number of reads that are duplicates;
}
\examples{
BamFilePaths <- dir(path="/home/ubuntu/Projects/RNAseqPackage/bams", pattern=".*.bam$", full.names=T)
UniqueReadsDF = SummarizeUniqueReads(BamFilePaths[1])
}
\author{
Nathaniel J. Madrid, Jason Byars
}
\keyword{SummarizeUniqueReads}
